# openapi.yaml
openapi: '3.0.0'
info:
  title: Priority ERP Mock API
  version: '1.0.0'
  description: Mock Priority ERP OData-style API for workshop on building AI agents with Google Gemini SDK
servers:
  - url: http://localhost:3000
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Order:
      type: object
      properties:
        ORDNAME:
          type: string
        CUSTNAME:
          type: string
        CURDATE:
          type: string
          format: date
        ORDSTATUS:
          type: string
          enum: [O, C, P]
        QPRICE:
          type: number
        DETAILS:
          type: string
        ORDERITEMS_SUBFORM:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    OrderItem:
      type: object
      properties:
        LINE:
          type: integer
          description: Auto-generated line number
        PARTNAME:
          type: string
        TQUANT:
          type: number
        PRICE:
          type: number
        DUEDATE:
          type: string
          format: date
        KLINE:
          type: string
...
    WorkOrder:
      type: object
      properties:
        WONUM:
          type: string
        PARTNAME:
          type: string
        TQUANT:
          type: number
        STARTDATE:
          type: string
          format: date
        ENDDATE:
          type: string
          format: date
        WOSTATUS:
          type: string
          enum: [P, A, C]
        PRIORITY:
          type: integer
    WorkOrderList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/WorkOrder'
...
  parameters:
    filterParam:
      name: $filter
      in: query
      description: OData filter expression
      schema:
        type: string
    orderbyParam:
      name: $orderby
      in: query
      description: OData orderby expression
      schema:
        type: string
    topParam:
      name: $top
      in: query
      description: Number of records to return
      schema:
        type: integer
    skipParam:
      name: $skip
      in: query
      description: Number of records to skip
      schema:
        type: integer
    selectParam:
      name: $select
      in: query
      description: Comma-separated list of fields to return
      schema:
        type: string
    expandParam:
      name: $expand
      in: query
      description: Subform to expand inline
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/ORDERS:
    get:
      summary: List orders
      operationId: listOrders
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
        - $ref: '#/components/parameters/expandParam'
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create an order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/ORDERS('{ORDNAME}')":
    get:
      summary: Get an order by key
      operationId: getOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update an order
      operationId: updateOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an order
      operationId: deleteOrder
      tags: [Orders]
      parameters:
        - name: ORDNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Order deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/WORKORDERS:
    get:
      summary: List work orders
      operationId: listWorkOrders
      tags: [Manufacturing]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of work orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a work order
      operationId: createWorkOrder
      tags: [Manufacturing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkOrder'
      responses:
        '201':
          description: Work order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/WORKORDERS('{WONUM}')":
    get:
      summary: Get a work order by key
      operationId: getWorkOrder
      tags: [Manufacturing]
      parameters:
        - name: WONUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Work order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a work order
      operationId: updateWorkOrder
      tags: [Manufacturing]
      parameters:
        - name: WONUM
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkOrder'
      responses:
        '200':
          description: Work order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a work order
      operationId: deleteWorkOrder
      tags: [Manufacturing]
      parameters:
        - name: WONUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Work order deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== BOM ====================
...